[REFACTOR]: исправил комментарии
